requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
PyJWT>=2.8.0
python-dotenv>=1.0.0
